Integer -> Long (versioned)

+++ fsl-server/src/main/java/lsfusion/erp/utils/UpdateProcessMonitorActionProperty.java
@@ -196,10 +196,10 @@ public class UpdateProcessMonitorActionProperty extends ScriptingActionProperty 
                 String fullQuery = (String) sqlProcess.get(2);
                 return fullQuery == null ? NullValue.instance : new DataObject(fullQuery);
             case "userProcess":
-                Integer user = (Integer) sqlProcess.get(3);
+                Long user = (Long) sqlProcess.get(3);
                 return user == null ? NullValue.instance : new DataObject(user);
             case "computerProcess":
-                Integer computer = (Integer) sqlProcess.get(4);
+                Long computer = (Long) sqlProcess.get(4);
                 return computer == null ? NullValue.instance : new DataObject(computer);
             case "addressUserSQLProcess":
                 String address = (String) sqlProcess.get(5);
@@ -438,8 +438,8 @@ public class UpdateProcessMonitorActionProperty extends ScriptingActionProperty 
                 Long startTransaction = sessionThread == null ? null : (Long) sessionThread.get(2);
                 String attemptCount = sessionThread == null ? "0" : (String) sessionThread.get(3);
                 StatusMessage statusMessage = sessionThread == null ? null : (StatusMessage) sessionThread.get(4);
-                Integer userActiveTask = sessionThread == null ? null : (Integer) sessionThread.get(5);
-                Integer computerActiveTask = sessionThread == null ? null : (Integer) sessionThread.get(6);
+                Long userActiveTask = sessionThread == null ? null : (Long) sessionThread.get(5);
+                Long computerActiveTask = sessionThread == null ? null : (Long) sessionThread.get(6);
                 String fullQuery = sessionThread == null || sessionThread.get(7) == null ? null : (String) sessionThread.get(7);
                 boolean isDisabledNestLoop = sessionThread != null && (boolean) sessionThread.get(8);
                 Integer queryTimeout = sessionThread == null ? null : (Integer) sessionThread.get(9);
